refactor: extract time-picker logic into reusable mixin
