+++ packages/time-picker/src/vaadin-time-picker.d.ts
@@ -4,17 +4,11 @@
  * This program is available under Apache License Version 2.0, available at https://vaadin.com/license/
  */
 import { ElementMixin } from '@vaadin/component-base/src/element-mixin.js';
-import { InputControlMixin } from '@vaadin/field-base/src/input-control-mixin.js';
-import { PatternMixin } from '@vaadin/field-base/src/pattern-mixin.js';
 import { ThemableMixin } from '@vaadin/vaadin-themable-mixin/vaadin-themable-mixin.js';
-import type { TimePickerTime } from './vaadin-time-picker-helper.js';
+import { TimePickerMixin } from './vaadin-time-picker-mixin.js';
 
-export type { TimePickerTime };
-
-export interface TimePickerI18n {
-  parseTime(time: string): TimePickerTime | undefined;
-  formatTime(time: TimePickerTime | undefined): string;
-}
+export type { TimePickerTime } from './vaadin-time-picker-helper.js';
+export type { TimePickerI18n } from './vaadin-time-picker-mixin.js';
 
 /**
  * Fired when the user commits a value change.
@@ -131,114 +125,7 @@ export interface TimePickerEventMap extends HTMLElementEventMap, TimePickerCusto
  * @fires {CustomEvent} value-changed - Fired when the `value` property changes.
  * @fires {CustomEvent} validated - Fired whenever the field is validated.
  */
-declare class TimePicker extends PatternMixin(InputControlMixin(ThemableMixin(ElementMixin(HTMLElement)))) {
-  /**
-   * The time value for this element.
-   *
-   * Supported time formats are in ISO 8601:
-   * - `hh:mm` (default)
-   * - `hh:mm:ss`
-   * - `hh:mm:ss.fff`
-   */
-  value: string;
-
-  /**
-   * True if the dropdown is open, false otherwise.
-   */
-  opened: boolean;
-
-  /**
-   * Minimum time allowed.
-   *
-   * Supported time formats are in ISO 8601:
-   * - `hh:mm`
-   * - `hh:mm:ss`
-   * - `hh:mm:ss.fff`
-   */
-  min: string;
-
-  /**
-   * Maximum time allowed.
-   *
-   * Supported time formats are in ISO 8601:
-   * - `hh:mm`
-   * - `hh:mm:ss`
-   * - `hh:mm:ss.fff`
-   */
-  max: string;
-
-  /**
-   * Specifies the number of valid intervals in a day used for
-   * configuring the items displayed in the selection box.
-   *
-   * It also configures the precision of the value string. By default
-   * the component formats values as `hh:mm` but setting a step value
-   * lower than one minute or one second, format resolution changes to
-   * `hh:mm:ss` and `hh:mm:ss.fff` respectively.
-   *
-   * Unit must be set in seconds, and for correctly configuring intervals
-   * in the dropdown, it need to evenly divide a day.
-   *
-   * Note: it is possible to define step that is dividing an hour in inexact
-   * fragments (i.e. 5760 seconds which equals 1 hour 36 minutes), but it is
-   * not recommended to use it for better UX experience.
-   */
-  step: number | null | undefined;
-
-  /**
-   * Set true to prevent the overlay from opening automatically.
-   * @attr {boolean} auto-open-disabled
-   */
-  autoOpenDisabled: boolean | null | undefined;
-
-  /**
-   * A space-delimited list of CSS class names to set on the overlay element.
-   *
-   * @attr {string} overlay-class
-   */
-  overlayClass: string;
-
-  /**
-   * The object used to localize this component.
-   * To change the default localization, replace the entire
-   * _i18n_ object or just the property you want to modify.
-   *
-   * The object has the following JSON structure:
-   *
-   * ```
-   * {
-   *   // A function to format given `Object` as
-   *   // time string. Object is in the format `{ hours: ..., minutes: ..., seconds: ..., milliseconds: ... }`
-   *   formatTime: (time) => {
-   *     // returns a string representation of the given
-   *     // object in `hh` / 'hh:mm' / 'hh:mm:ss' / 'hh:mm:ss.fff' - formats
-   *   },
-   *
-   *   // A function to parse the given text to an `Object` in the format
-   *   // `{ hours: ..., minutes: ..., seconds: ..., milliseconds: ... }`.
-   *   // Must properly parse (at least) text
-   *   // formatted by `formatTime`.
-   *   parseTime: text => {
-   *     // Parses a string in object/string that can be formatted by`formatTime`.
-   *   }
-   * }
-   * ```
-   *
-   * Both `formatTime` and `parseTime` need to be implemented
-   * to ensure the component works properly.
-   */
-  i18n: TimePickerI18n;
-
-  /**
-   * Opens the dropdown list.
-   */
-  open(): void;
-
-  /**
-   * Closes the dropdown list.
-   */
-  close(): void;
-
+declare class TimePicker extends TimePickerMixin(ThemableMixin(ElementMixin(HTMLElement))) {
   addEventListener<K extends keyof TimePickerEventMap>(
     type: K,
     listener: (this: TimePicker, ev: TimePickerEventMap[K]) => void,
